Add files via upload
to add the camera create and empty game object for the camera. Create a UI for it under the ui canvas as UI child images,  photodisplay and camerabutton, drag the sprites for the camera button into their sprite renderers. Create the camera script and attach to the Camera empty game object then drag the premade images you want to display into the images array folder, then go to the camera button, go to on click and then drag the amera game object into it, go to the functions tab next to it and go to camera script then on the drop-down menu choose take fake photo.